// components/home_components/interfaces.tsx
export interface ChatGroupDataType {
  groupTitle: string;
  salonsTitle: string[];
  salons: string[][];
  groupIconSrc: string;
  totalMember: number;
  totalMemberOnline: number;
}
export interface UserContact {
  profileImgSrc: string;
  firstName: string;
  lastName: string;
  color: string;
  online: boolean;
}